requirements.txt
# Discord Bot
discord.py==2.6.4

# MQTT Client
paho-mqtt==2.1.0

# Environment Variables
python-dotenv==1.2.1
